KeyValueStore 및 ServiceLeaderLock 엔티티 갱신 시 updateAt 수정하기.
KeyValueStoreRepository 에서 id 타입을 Long 으로 변경.
ServiceLeaderLock 에서 id 타입을 Long 으로 변경.

+++ src/main/kotlin/com/kimpscan/api/exchange/repository/KeyValueStoreRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Lock
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 
-interface KeyValueStoreRepository : JpaRepository<KeyValueStore, Int> {
+interface KeyValueStoreRepository : JpaRepository<KeyValueStore, Long> {
 
     @Lock(LockModeType.PESSIMISTIC_WRITE)
     @Query("SELECT k FROM KeyValueStore k WHERE k.key = :key")
